Re-did home_widget.dart -> home_widget2.dart and Dropdown works
