+++ lib/placeholder_widget.dart
@@ -2,8 +2,12 @@ import 'package:flutter/material.dart';
 
 class PlaceholderWidget extends StatelessWidget{
   final Color color;
-  PlaceholderWidget(this.color);
+  PlaceholderWidget(this.color,);
 
   @override
-  Widget build(BuildContext context) => Container(color: color,);
+  Widget build(BuildContext context) {
+    return Container(
+      color: color,
+    );
+  }
 }
